refactor: move execute() up to AbstractMetaSyncCommand
